make camera move
2d 카메라 이동

+++ New/Assets/Scripts/Tower/TowerStatus.cs
@@ -2,13 +2,12 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-
-
 public class TowerStatus : MonoBehaviour
 {
     public string element;
     public string type;
-    public float attack;
+
+    public int attack;
     public float attack_speed;
     public int direction;
 
@@ -16,5 +15,4 @@ public class TowerStatus : MonoBehaviour
 
     // 상점
     //int cost = 1;
-
 }
